pyproject.toml
[project]
name = "agent-ai"
version = "0.1.0"
description = "✨ Futuristic Urdu Translator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "google-cloud-translate>=3.20.3",
    "litellm>=1.72.4",
    "openai>=1.86.0",
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
]
